functions/GreetingFunction/requirements.txt
crowdstrike-foundry-function==1.1.3